backend/printer-control/requirements.txt
# Printer Control Dependencies
# Python 3.12+

# Core dependencies
bambulabs-api>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.0.0
structlog>=24.0.0
paho-mqtt>=1.6.0
pyyaml>=6.0.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
ruff>=0.7.0
mypy>=1.7.0
types-pyyaml>=6.0.0
